job_filling/requirements.txt
pymupdf
pytest
playwright
anthropic
pyyaml
